+++ src/Symfony/Component/Validator/ConstraintViolation.php
@@ -56,6 +56,10 @@ class ConstraintViolation implements ConstraintViolationInterface
             $message = '';
         }
 
+        if (null !== $code && !\is_string($code)) {
+            @trigger_error(sprintf('Not using a string as the error code in %s() is deprecated since Symfony 4.4. A type-hint will be added in 5.0.', __METHOD__), E_USER_DEPRECATED);
+        }
+
         $this->message = $message;
         $this->messageTemplate = $messageTemplate;
         $this->parameters = $parameters;