+++ src/Symfony/Component/Validator/Tests/ConstraintViolationTest.php
@@ -64,7 +64,7 @@ EOF;
             'some_value',
             null,
             null,
-            0
+            '0'
         );
 
         $expected = <<<'EOF'
@@ -107,5 +107,25 @@ EOF;
         );
 
         $this->assertSame($expected, (string) $violation);
+    }
+
+    /**
+     * @group legacy
+     * @expectedDeprecation Not using a string as the error code in Symfony\Component\Validator\ConstraintViolation::__construct() is deprecated since Symfony 4.4. A type-hint will be added in 5.0.
+     */
+    public function testNonStringCode()
+    {
+        $violation = new ConstraintViolation(
+            '42 cannot be used here',
+            'this is the message template',
+            [],
+            ['some_value' => 42],
+            'some_value',
+            null,
+            null,
+            42
+        );
+
+        self::assertSame(42, $violation->getCode());
     }
 }